Swap Tokens functionality!
Add Balances > Swap Tokens dialog, which allows to atomically and
directly exchange tokens with a partner.
It also provides an easy way for exchanging symetric informations (to
Taker), through an weakly check-summed string, that can be copied and
sent to Taker, who can then fill all informations only by pasting this
string in the proper field.